rename msg_util to msg_utils
